hw1/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
numpy
pandas
tqdm
pillow
pyyaml
tensorboard
albumentations
opencv-python
cleanlab
cleanvision
matplotlib
seaborn
scikit-learn
